Handle the case where foreign arguments are inserted before our own
AFAICT,	a CommandLineTool's `arguments` will always be contiguous, but
they won't necessarily be at the start of the command line. If another
argument's `position` is negative (all arguments default to position 0),
that argument will appear before any arguments with unspecified
position. Therefore, we now wait until we see a start token before
starting to create directories.

Unfortunately expressions are not allowed in baseCommand, so we can't
just put everything in there.