Ignore declared type of outputs
Any output can have a glob regardless of type, as long as it has an
outputEval that turns the matched array of files into the output type
that was declared. If we detect that an output is matching against a
glob, we should create a file (or multiple files) that match that glob,
even if the output itself isn't a file/array of files.